backend/requirements.txt
# Django & REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Redis & Celery
redis==5.0.1
celery==5.3.4

# File handling & Document processing
Pillow==10.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract==0.3.10
openai==1.3.5
httpx==0.27.0  # Compatible with openai 1.3.5
google-generativeai==0.3.2  # Google Gemini API (free alternative)
reportlab==4.0.7  # For PDF generation
requests==2.31.0  # For downloading files from URL

# AWS S3 (optional cloud storage)
boto3==1.29.7
django-storages==1.14.2

# Environment variables
python-dotenv==1.0.0

# CORS
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# Production server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.11.0
flake8==6.1.0

